Fixes #5204 - SNI does not work with PKIX.
Updates after review.
